requirements.txt
robotframework==2.9.2
robotframework-selenium2library==1.7.4
robotframework-debuglibrary==0.4
Selenium==2.53.1
factory_boy==2.6.1
requests==2.9.1
pytest==2.9.1
pytest-django==2.9.1
psycopg2==2.6.1
mock==2.0.0

